pyproject.toml
[tool.poetry]
name = "stock-tracker"
version = "0.1.0"
description = ""
authors = ["Brent Gruber <brent.gruber77@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
praw = "^7.1.4"
APScheduler = "^3.7.0"
requests-async = "^0.6.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.16.21"}
aerich = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
